Fix some things
