+++ test/tests/logging_tests.py
@@ -148,77 +148,3 @@ class LoggingTests(PavTestCase):
 
         self.assertIn(ident, handler.ERR_OUT.getvalue())
         self.assertNotIn(ident, logfile_path.open().read())
-
-    def test_series_file_logger(self):
-        """Test that the series file logger works correctly."""
-
-        arg_parser = arguments.get_parser()
-
-        args = arg_parser.parse_args([
-            'run',
-            '-H', 'this',
-            'results_log',
-        ])
-
-        run_cmd = commands.get_command(args.command_name)
-        run_cmd.silence()
-
-        log_path = self.pav_cfg.working_dir / "results"
-        self.pav_cfg = self.make_pav_config(result_loggers= [{
-                                                "plugin": "series_file",
-                                                "dest": str(log_path)}])
-
-        self.assertEqual(run_cmd.run(self.pav_cfg, args), 0)
-
-        series = run_cmd.last_series
-        series.wait_log(timeout=10)
-
-        matches = list(log_path.glob(f"{series.id}*"))
-
-        self.assertEqual(len(matches), 1,
-                         msg=f"Expected exactly one log file matching '{series.id}*', "
-                             f"but found {len(matches)}: {matches}")
-
-        log_path = next(iter(matches))
-
-        with open(log_path) as fin:
-            results = json.load(fin)
-
-        self.assertEqual(results.get("hello"), "world")
-
-    def test_common_file_logger(self):
-        """Test that the common file logger works correctly."""
-
-        arg_parser = arguments.get_parser()
-
-        args = arg_parser.parse_args([
-            'run',
-            '-H', 'this',
-            'results_log',
-        ])
-
-        run_cmd = commands.get_command(args.command_name)
-        run_cmd.silence()
-
-        log_path = self.pav_cfg.working_dir / "results.log"
-        self.pav_cfg = self.make_pav_config(result_loggers=[{
-                                                "plugin": "common_file",
-                                                "dest": str(log_path)}])
-
-        self.assertEqual(run_cmd.run(self.pav_cfg, args), 0)
-        series1 = run_cmd.last_series
-
-        self.assertEqual(run_cmd.run(self.pav_cfg, args), 0)
-        series2 = run_cmd.last_series
-
-        series1.wait_log(timeout=10)
-        series2.wait_log(timeout=10)
-
-        with open(log_path) as fin:
-            results = fin.readlines()
-
-        self.assertEqual(len(results), 2)
-
-        for res in results:
-            results = json.loads(res)
-            self.assertEqual(results.get("hello"), "world")
